Rename `cleanup` module to `housekeeping`
