+++ daemon/src/housekeeping.rs
